mysite-house/requirements.txt
certifi==2021.10.8
Django==2.2.5
django-admin-tools==0.9.2
django-filter==21.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-tables2==2.4.1
Pillow==9.0.0
PyMySQL==1.0.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2